codematrix_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
google-generativeai
GitPython
aiofiles 